cdp1802 testsuite: use long where appropriate on 16-bit targets
Avoids test failures due to extraneous warnings on 16-bit targets.

gcc/testsuite/

	* c-c++-common/Warray-bounds-3.c: Use long for 32-bit field on int16.
	* gcc.dg/analyzer/torture/switch-4.c: Use unsigned long for id field.
	* gcc.dg/cpp/ucs.c (foo): Use long for c.
	* gcc.dg/pr115646.c: Use long for bit fields wider than 16.
	* gcc.dg/pr93231.c: Use long for expression shifted more than 16 bits.

+++ gcc/testsuite/gcc.dg/pr115646.c
@@ -5,7 +5,7 @@
 extern double pow(double x, double y);
 
 struct S {
-    unsigned int a : 3, b : 8, c : 21;
+    unsigned long a : 3, b : 8, c : 21;
 };
 
 void foo (struct S *p)
